Refactored code to use pcq-commons library
