+++ build.gradle
@@ -268,6 +268,7 @@ dependencies {
     compile group: 'uk.gov.hmcts.reform', name: 'core-case-data-store-client', version: '4.7.3'
 
     implementation group: 'org.springframework.boot', name: 'spring-boot-starter-validation'
+    implementation group: 'uk.gov.hmcts.reform.pcq', name: 'pcq-commons', version: '1.1.2'
 
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: versions.junit
     testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-params', version: versions.junit
