+++ src/functionalTest/java/uk/gov/hmcts/reform/pcqconsolidationservice/postdeploy/ConsolidationServiceFunctionalTest.java
@@ -11,9 +11,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.util.ReflectionUtils;
+import uk.gov.hmcts.reform.pcq.commons.model.PcqAnswerResponse;
 import uk.gov.hmcts.reform.pcqconsolidationservice.ConsolidationComponent;
 import uk.gov.hmcts.reform.pcqconsolidationservice.config.TestApplicationConfiguration;
-import uk.gov.hmcts.reform.pcqconsolidationservice.controller.response.PcqAnswerResponse;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
