+++ src/integrationTest/java/uk/gov/hmcts/reform/pcqconsolidationservice/util/SpringBootIntegrationTest.java
@@ -4,9 +4,9 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
+import uk.gov.hmcts.reform.pcq.commons.controller.feign.PcqBackendFeignClient;
 import uk.gov.hmcts.reform.pcqconsolidationservice.ConsolidationComponent;
 import uk.gov.hmcts.reform.pcqconsolidationservice.config.TestApplicationConfiguration;
-import uk.gov.hmcts.reform.pcqconsolidationservice.controller.feign.PcqBackendFeignClient;
 import uk.gov.hmcts.reform.pcqconsolidationservice.service.impl.PcqBackendServiceImpl;
