+++ src/main/java/uk/gov/hmcts/reform/pcqconsolidationservice/ConsolidationApplication.java
@@ -14,7 +14,7 @@ import org.springframework.context.ApplicationContext;
 
 @SpringBootApplication(scanBasePackages = "uk.gov.hmcts.reform")
 @EnableFeignClients(basePackages = {
-        "uk.gov.hmcts.reform.pcqconsolidationservice"
+        "uk.gov.hmcts.reform.pcq.commons"
 })
 @Slf4j
 public class ConsolidationApplication implements ApplicationRunner {
