--- src/main/java/uk/gov/hmcts/reform/pcqconsolidationservice/config/FeignHeaderConfig.java
@@ -1,16 +1,0 @@
-package uk.gov.hmcts.reform.pcqconsolidationservice.config;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
-
-@Configuration
-@ConfigurationProperties(prefix = "feign.allow")
-@Getter
-@AllArgsConstructor
-public class FeignHeaderConfig {
-    private final List<String> headers;
-}
