--- src/main/java/uk/gov/hmcts/reform/pcqconsolidationservice/config/FeignInterceptorConfiguration.java
@@ -1,40 +1,0 @@
-package uk.gov.hmcts.reform.pcqconsolidationservice.config;
-
-import feign.RequestInterceptor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-import java.util.Enumeration;
-import java.util.Locale;
-import javax.servlet.http.HttpServletRequest;
-
-
-@Slf4j
-@SuppressWarnings("PMD.ConfusingTernary")
-public class FeignInterceptorConfiguration {
-
-    @Bean
-    public RequestInterceptor requestInterceptor(FeignHeaderConfig config) {
-        return requestTemplate -> {
-            ServletRequestAttributes attrs = (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();
-            if (attrs != null) {
-                HttpServletRequest request = attrs.getRequest();
-                Enumeration<String> headerNames = request.getHeaderNames();
-                if (headerNames != null) {
-                    while (headerNames.hasMoreElements()) {
-                        String name = headerNames.nextElement();
-                        String value = request.getHeader(name);
-                        if (config.getHeaders().contains(name.toLowerCase(Locale.UK))) {
-                            requestTemplate.header(name, value);
-                        }
-                    }
-                } else {
-                    log.warn("FeignHeadConfiguration", "Failed to get request header!");
-                }
-            }
-        };
-    }
-
-}
